refactoring of the studentday class
made the attendant, notAttendant, notSetAtt  static because they are the same for all instances of the class.
made the timeDate string into a localdate, so that all the classes in the program use the same type of date class.
added an attendance status integer to the Students Day.
added a student variable to the students day, so that it is always clear for which student the day is describing the attendance for.
added a Constructor for the studentDay.
added getters and setters for all the added variables.