teacherStudentList and currentteacher and currentclass in modelfacade.
teacherStudentList and currentteacher and currentclass in modelfacade.
getallstudents for class in DAL implemented.
getallclasses for teacher and getallstudents for class work in gui.
you can also click on a class as a teacher and get alle the students that class. You can also click on a student and go to their page and se theirs stats and change the attendance.